+++ pyslvs_ui/entities/edit_point.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""The option dialog use to create or edit the point."""
+"""The option dialog used to create or edit the point."""
 
 __author__ = "Yuan Chang"
 __copyright__ = "Copyright (C) 2016-2020"
@@ -13,6 +13,7 @@ from qtpy.QtWidgets import QDialog, QListWidgetItem, QColorDialog, QWidget
 from qtpy.QtGui import QIcon, QPixmap
 from pyslvs import VPoint, VLink
 from pyslvs_ui.graphics import color_names, color_qt, color_icon
+from .utility import set_custom_color, add_custom_color
 from .edit_point_ui import Ui_Dialog
 
 
@@ -25,7 +26,7 @@ class EditPointDialog(QDialog, Ui_Dialog):
         vpoints: List[VPoint],
         vlinks: List[VLink],
         pos: Union[int, bool],
-        parent: QWidget,
+        parent: QWidget
     ):
         """Input data reference from main window.
 
@@ -44,7 +45,7 @@ class EditPointDialog(QDialog, Ui_Dialog):
         for i, e in enumerate(color_names):
             self.color_box.insertItem(i, color_icon(e), e)
         for vlink in vlinks:
-            self.noSelected.addItem(QListWidgetItem(self.link_icon, vlink.name))
+            self.no_selected.addItem(QListWidgetItem(self.link_icon, vlink.name))
         if pos is False:
             self.name_box.addItem(icon, f'Point{vpoints_count}')
             self.name_box.setEnabled(False)
@@ -53,6 +54,8 @@ class EditPointDialog(QDialog, Ui_Dialog):
             for i in range(vpoints_count):
                 self.name_box.insertItem(i, icon, f'Point{i}')
             self.name_box.setCurrentIndex(pos)
+        self.type_box.currentIndexChanged.connect(self.__check_angle)
+        self.__check_angle()
 
     @Slot(int, name='on_name_box_currentIndexChanged')
     def __set_name(self, index: int) -> None:
@@ -62,23 +65,17 @@ class EditPointDialog(QDialog, Ui_Dialog):
         vpoint = self.vpoints[index]
         self.x_box.setValue(vpoint.x)
         self.y_box.setValue(vpoint.y)
-        color_text = vpoint.color_str
-        color_index = self.color_box.findText(color_text)
-        if color_index > -1:
-            self.color_box.setCurrentIndex(color_index)
-        else:
-            self.color_box.addItem(color_icon(color_text), color_text)
-            self.color_box.setCurrentIndex(self.color_box.count() - 1)
+        set_custom_color(self.color_box, vpoint.color_str)
         self.type_box.setCurrentIndex(vpoint.type)
         self.angle_box.setValue(vpoint.angle)
-        self.noSelected.clear()
+        self.no_selected.clear()
         self.selected.clear()
         for linkName in vpoint.links:
             self.selected.addItem(QListWidgetItem(self.link_icon, linkName))
         for vlink in self.vlinks:
             if vlink.name in vpoint.links:
                 continue
-            self.noSelected.addItem(QListWidgetItem(self.link_icon, vlink.name))
+            self.no_selected.addItem(QListWidgetItem(self.link_icon, vlink.name))
 
     @Slot(int, name='on_color_box_currentIndexChanged')
     def __set_color(self, _=None) -> None:
@@ -94,27 +91,26 @@ class EditPointDialog(QDialog, Ui_Dialog):
             color_qt(self.color_box.currentText()),
             self
         )
-        if not color.isValid():
-            return
-        rgb_str = str((color.red(), color.green(), color.blue()))
-        self.color_box.addItem(color_icon(rgb_str), rgb_str)
-        self.color_box.setCurrentIndex(self.color_box.count() - 1)
+        if color.isValid():
+            add_custom_color(self.color_box, color)
 
-    @Slot(int, name='on_type_box_currentIndexChanged')
-    def __set_type(self, index: int) -> None:
+    @Slot()
+    def __check_angle(self) -> None:
         """Toggle the slider angle option."""
-        self.angle_box.setEnabled(index != 0)
+        enabled = self.type_box.currentIndex() != 0
+        self.angle_label.setEnabled(enabled)
+        self.angle_box.setEnabled(enabled)
 
-    @Slot(QListWidgetItem, name='on_noSelected_itemDoubleClicked')
+    @Slot(QListWidgetItem, name='on_no_selected_itemDoubleClicked')
     def __add_selected(self, item: QListWidgetItem) -> None:
         """Add item to selected list."""
         self.selected.addItem(
-            self.noSelected.takeItem(self.noSelected.row(item))
+            self.no_selected.takeItem(self.no_selected.row(item))
         )
 
     @Slot(QListWidgetItem, name='on_selected_itemDoubleClicked')
     def __add_no_selected(self, item: QListWidgetItem) -> None:
         """Add item to no selected list."""
-        self.noSelected.addItem(
+        self.no_selected.addItem(
             self.selected.takeItem(self.selected.row(item))
         )