+++ pyslvs_ui/widgets/main_base.py
@@ -337,7 +337,7 @@ class MainWindowBase(MainWindowABC, ABC):
         clean_selection_action.triggered.connect(table_clear_selection)
         clean_selection_action.setShortcut("Esc")
         clean_selection_action.setShortcutContext(Qt.WindowShortcut)
-        self.main_canvas.noselected.connect(table_clear_selection)
+        self.main_canvas.no_selected.connect(table_clear_selection)
         self.addAction(clean_selection_action)
 
         self.main_canvas.free_moved.connect(self.set_free_move)
@@ -373,7 +373,7 @@ class MainWindowBase(MainWindowABC, ABC):
                 self.inputs_widget.set_selection(selections)
 
         self.main_canvas.selected.connect(inputs_selection)
-        self.main_canvas.noselected.connect(self.inputs_widget.clear_selection)
+        self.main_canvas.no_selected.connect(self.inputs_widget.clear_selection)
         self.inputs_widget.update_preview_button.clicked.connect(self.main_canvas.update_preview_path)
 
         # Synthesis collections